+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Next release
+## v0.2.0
 
 - Implement frontend building on deploy
 - Implement deployment by docker compose